+++ euler.h
@@ -1,6 +1,6 @@
-int a(int p, int q);
-int b(int p, int q);
-int c(int p, int q);
-void print_triples(long *triples);
-long* cons_triple(int a, int b, int c);
-long* cons_duo(int a, int b);
+long long a(long long p, long long q);
+long long b(long long p, long long q);
+long long c(long long p, long long q);
+void print_triples(long long *triples);
+long long* cons_triple(long long a, long long b, long long c);
+long long* cons_duo(long long a, long long b);
